requirements.txt
openai
pandas
numpy
flask
python-dotenv
mysql.connector
pytesseract
opencv-python